+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Version 0.2.0
 + Add cache for rendering markdown
 + Add cache for ``url_for``
 + Add cache for ``static_url``
++ Use Jinja2's Bytecode Cache
 + Don't generate analytics codes when running ``catsup server``
 + Display time cost for loading config and posts